backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25

alembic==1.13.1
cloudscraper==1.2.71
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.0
langchain-google-genai==2.0.9
google-generativeai==0.8.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.3
